requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
camelot-py==1.0.9
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.127.0
fire==0.7.1
fonttools==4.61.1
h11==0.16.0
idna==3.11
lxml==6.0.2
numpy==2.2.6
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
pandas==2.3.3
pdf2docx==0.5.8
pdf2image==1.17.0
pdfminer.six==20251107
pillow==12.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.26.7
pypdf==3.17.4
pypdfium2==5.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-multipart==0.0.21
python-pptx==1.0.2
pytz==2025.2
six==1.17.0
starlette==0.50.0
tabulate==0.9.0
termcolor==3.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.40.0
xlrd==2.0.2
xlsxwriter==3.2.9
xlwt==1.3.0
